blocktorch/experimental/requirements_lock.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
datasets==3.6.0
dill==0.3.8
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
ml_dtypes==0.5.1
mlx==0.26.1
mpmath==1.3.0
multidict==6.6.2
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
-e git+ssh://git@github.com/Ian-Reitsma/Apple-Metal-Orchard.git@f7c7e1451055473926f97d108aa92fa17ffffaff#egg=orchard
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
psutil==7.0.0
pyarrow==20.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools-scm==8.3.1
six==1.17.0
sympy==1.14.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
